Rename network utils file
